+++ src/Web/ComponentUploadField.ts
@@ -11,7 +11,7 @@ export class ComponentUploadField extends HTMLElement {
     this.innerHTML = `\
 <div>
   <label for="upload">Import file</label>
-  <input type="file" id="upload" accept=".mcf,.tmc,.fpl,.pln,.fms,.json,.gpx,.geojson" />
+  <input type="file" id="upload" accept=".mcf,.tmc,.fpl,.pln,.fms,.json,.gpx,.geojson,.cup" />
 </div>
 `;